requirements.txt
streamlit
numpy
Pillow
tensorflow
keras